requirements.txt
beautifulsoup4==4.9.3
requests==2.26.0
selenium==3.141.0
webdriver_manager==3.4.2
python-telegram-bot==20.5
arrow==1.2.3